refactor(gateway): move types to separate file
