+++ src/controllers/GatewayController.ts
@@ -1,9 +1,11 @@
 import { inject, singleton } from 'tsyringe';
-import GatewayService, { GatewayPacket, GatewayPacketType, AuthenticateGatewayPacket, HelloGatewayPacket } from '../services/GatewayService';
+import GatewayService from '../services/GatewayService';
 import WebSocket, { Server as WebSocketServer, Data } from 'ws';
 import { User } from '../entities/User';
 import { UserService } from '../services/UserService';
 import { verifyJWT } from '../util/auth';
+import { GatewayPacket, GatewayPacketType, HelloGatewayPacket, IdentifyGatewayPacket, GatewayError } from '../util/gateway';
+import { getConfig } from '../util/config';
 
 @singleton()
 export default class GatewayController {
@@ -19,16 +21,17 @@ export default class GatewayController {
 	}
 
 	public bindTo(wss: WebSocketServer) {
-		if (this.wss) throw new Error('WebSocketServer is already bound!');
+		if (this.wss) throw new GatewayError('WebSocketServer is already bound!');
 		this.wss = wss;
 		this.wss.on('connection', ws => {
 			ws.on('message', data => void this.onMessage(ws, data).catch(error => {
-				console.error(error);
+				if (!(error instanceof GatewayError) && getConfig().logErrors) {
+					console.error(error);
+				}
 				ws.close();
 			}));
-			ws.on('close', (code, reason) => {
+			ws.on('close', () => {
 				this.authenticatedClients.delete(ws);
-				console.log(code, reason);
 			});
 		});
 	}
@@ -39,24 +42,25 @@ export default class GatewayController {
 	}
 
 	private async onPacket(ws: WebSocket, packet: GatewayPacket) {
-		if (packet.t === GatewayPacketType.Authenticate) {
-			await this.onAuthenticate(ws, packet as AuthenticateGatewayPacket);
+		switch (packet.t) {
+			case GatewayPacketType.Identify:
+				await this.onAuthenticate(ws, packet as IdentifyGatewayPacket);
+				break;
+			default:
+				throw new GatewayError(`Received invalid packet type ${packet.t}`);
 		}
 	}
 
-	public async onAuthenticate(ws: WebSocket, packet: AuthenticateGatewayPacket) {
+	public async onAuthenticate(ws: WebSocket, packet: IdentifyGatewayPacket) {
 		let user = this.authenticatedClients.get(ws);
-		if (user) throw new Error('Already authenticated!');
+		if (user) throw new GatewayError('Already authenticated!');
 		const { token } = packet.d;
 		const { id } = await verifyJWT(token);
 		user = await this.userService.findOne({ id });
-		if (!user) throw new Error('User not found');
+		if (!user) throw new GatewayError('User not found');
 		this.authenticatedClients.set(ws, user);
 		await this.gatewayService.send([ws], {
-			t: GatewayPacketType.Hello,
-			d: {
-				time: new Date().toISOString()
-			}
+			t: GatewayPacketType.Hello
 		} as HelloGatewayPacket);
 		return user;
 	}
